// packages/apps/frontend/src/app/features/post-list/generated/graphql.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** Cursor for paging through collections */
  ConnectionCursor: { input: any; output: any; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
};

/** Information about the application */
export type AppInfo = {
  __typename?: 'AppInfo';
  /** A brief description of the application */
  description: Scalars['String']['output'];
  /** The name of the application */
  name: Scalars['String']['output'];
  /** The title of the application */
  title: Scalars['String']['output'];
  /** The current version of the application */
  version: Scalars['String']['output'];
};

export type BooleanFieldComparison = {
  is?: InputMaybe<Scalars['Boolean']['input']>;
  isNot?: InputMaybe<Scalars['Boolean']['input']>;
};

/** Comment model */
export type Comment = {
  __typename?: 'Comment';
  content: Scalars['String']['output'];
  /** Timestamp of resource creation */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates if the resource is deleted */
  deleted?: Maybe<Scalars['Boolean']['output']>;
  /** Timestamp of when the resource was deleted */
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Unique identifier for the resource */
  id: Scalars['ID']['output'];
  /** Post associated with the comment */
  post: Post;
  /** Timestamp of last update to the resource */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};

export type CommentConnection = {
  __typename?: 'CommentConnection';
  /** Array of edges. */
  edges: Array<CommentEdge>;
  /** Paging information */
  pageInfo: PageInfo;
  /** Fetch total count of records */
  totalCount: Scalars['Int']['output'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** The cursor of the last returned record. */
  endCursor?: Maybe<Scalars['ConnectionCursor']['output']>;
  /** true if paging forward and there are more records. */
  hasNextPage?: Maybe<Scalars['Boolean']['output']>;
  /** true if paging backwards and there are more records. */
  hasPreviousPage?: Maybe<Scalars['Boolean']['output']>;
  /** The cursor of the first returned record. */
  startCursor?: Maybe<Scalars['ConnectionCursor']['output']>;
};

/** Post model */
export type Post = {
  __typename?: 'Post';
  /** Comments associated with the post */
  comments?: Maybe<PostCommentsConnection>;
  /** Content of the post */
  content: Scalars['String']['output'];
  /** Timestamp of resource creation */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates if the resource is deleted */
  deleted?: Maybe<Scalars['Boolean']['output']>;
  /** Timestamp of when the resource was deleted */
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Unique identifier for the resource */
  id: Scalars['ID']['output'];
  /** Sources of the post */
  sources?: Maybe<Array<Scalars['String']['output']>>;
  /** Teaser of the post */
  teaser: Scalars['String']['output'];
  /** Title of the post */
  title: Scalars['String']['output'];
  /** Timestamp of last update to the resource */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type PostConnection = {
  __typename?: 'PostConnection';
  /** Array of edges. */
  edges: Array<PostEdge>;
  /** Paging information */
  pageInfo: PageInfo;
  /** Fetch total count of records */
  totalCount: Scalars['Int']['output'];
};
// ...
export type PostEdge = {
  __typename?: 'PostEdge';
  /** Cursor for this node. */
  cursor: Scalars['ConnectionCursor']['output'];
  /** The node containing the Post */
  node: Post;
};
// ...
export type PostSort = {
  direction: SortDirection;
  field: PostSortFields;
  nulls?: InputMaybe<SortNulls>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Get application information */
  appInfo: AppInfo;
  comment: Comment;
  comments: CommentConnection;
  /** Get the current user */
  me: User;
  post: Post;
  posts: PostConnection;
  role: Role;
  roles: RoleConnection;
  /** Get a list of all scheduled tasks */
  tasks: Array<TaskInfo>;
  user: User;
  users: UserConnection;
};
// ...
export type Role = {
  __typename?: 'Role';
  /** Timestamp of resource creation */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates if the resource is deleted */
  deleted?: Maybe<Scalars['Boolean']['output']>;
  /** Timestamp of when the resource was deleted */
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Description of the role */
  description?: Maybe<Scalars['String']['output']>;
  /** Unique identifier for the resource */
  id: Scalars['ID']['output'];
  /** Indicates if the role is a system role */
  isSystemRole: Scalars['Boolean']['output'];
  /** Name of the role */
  name: Scalars['String']['output'];
  /** Timestamp of last update to the resource */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Users associated with the role */
  users?: Maybe<RoleUsersConnection>;
};


/** Role model */
export type RoleUsersArgs = {
  filter?: UserFilter;
  paging?: CursorPaging;
  sorting?: Array<UserSort>;
};

export type RoleConnection = {
  __typename?: 'RoleConnection';
  /** Array of edges. */
  edges: Array<RoleEdge>;
  /** Paging information */
  pageInfo: PageInfo;
  /** Fetch total count of records */
  totalCount: Scalars['Int']['output'];
};
// ...
export type TaskInfo = {
  __typename?: 'TaskInfo';
  /** The last execution time of the task */
  lastDate?: Maybe<Scalars['DateTime']['output']>;
  /** The result of the last execution */
  lastResult?: Maybe<TaskResult>;
  /** The name of the task */
  name: Scalars['String']['output'];
  /** The next scheduled execution time of the task */
  nextDate: Scalars['DateTime']['output'];
  /** The current state of the task */
  state: TaskState;
};
// ...
export type User = {
  __typename?: 'User';
  /** Timestamp of resource creation */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates if the resource is deleted */
  deleted?: Maybe<Scalars['Boolean']['output']>;
  /** Timestamp of when the resource was deleted */
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Email address of the user */
  email: Scalars['String']['output'];
  /** Unique identifier for the resource */
  id: Scalars['ID']['output'];
  /** Timestamp of the user's last login */
  lastLogin?: Maybe<Scalars['DateTime']['output']>;
  /** Role of the user */
  role?: Maybe<Role>;
  /** Timestamp of last update to the resource */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Username of the user */
  username: Scalars['String']['output'];
};

export type UserConnection = {
  __typename?: 'UserConnection';
  /** Array of edges. */
  edges: Array<UserEdge>;
  /** Paging information */
  pageInfo: PageInfo;
  /** Fetch total count of records */
  totalCount: Scalars['Int']['output'];
};
// ...
export type GetPostsQueryVariables = Exact<{
  first?: InputMaybe<Scalars['Int']['input']>;
  after?: InputMaybe<Scalars['ConnectionCursor']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
  before?: InputMaybe<Scalars['ConnectionCursor']['input']>;
  sorting: Array<PostSort> | PostSort;
}>;


export type GetPostsQuery = { __typename?: 'Query', posts: { __typename?: 'PostConnection', totalCount: number, edges: Array<{ __typename?: 'PostEdge', node: { __typename?: 'Post', id: string, title: string, teaser: string, createdAt?: any | null } }>, pageInfo: { __typename?: 'PageInfo', hasNextPage?: boolean | null, hasPreviousPage?: boolean | null, startCursor?: any | null, endCursor?: any | null } } };

export const GetPostsDocument = gql`
    query GetPosts($first: Int, $after: ConnectionCursor, $last: Int, $before: ConnectionCursor, $sorting: [PostSort!]!) {
  posts(
    paging: {first: $first, after: $after, last: $last, before: $before}
    sorting: $sorting
  ) {
    totalCount
    edges {
      node {
        id
        title
        teaser
        createdAt
      }
    }
    pageInfo {
      hasNextPage
      hasPreviousPage
      startCursor
      endCursor
    }
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class GetPostsGQL extends Apollo.Query<GetPostsQuery, GetPostsQueryVariables> {
    override document = GetPostsDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }